backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pymysql==1.1.1
bcrypt==4.2.0
python-dotenv==1.0.1
PyJWT==2.9.0
